Refactor Rust core integration by migrating to a new native module. This update includes the introduction of a `memori.native` package, consolidating Rust-related functionality and improving the organization of the codebase. The legacy `memori._rust_core` imports have been replaced with the new structure across various modules, enhancing maintainability and clarity. Additionally, unit tests have been updated to reflect these changes, ensuring compatibility with the new architecture.
