+++ memori/native/_errors.py
@@ -1,0 +1,5 @@
+"""Shared exceptions for the native Rust core bridge."""
+
+
+class RustCoreAdapterError(RuntimeError):
+    pass
